backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3